Restructure the project data layer
