+++ src/app/core/data-layer/activity.reducer.ts
