+++ src/app/core/data-layer/devices.reducer.ts
