+++ src/app/core/data-layer/widgets.reducer.ts
