+++ src/app/core/services/communicate.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed, inject } from '@angular/core/testing';
 import { StoreModule } from '@ngrx/store';
-import { appReducersGenerator } from '@app/core/app.reducers';
+import { appReducersGenerator } from '@app/app.reducers';
 import { CommunicateService } from './communicate.service';
 import { RequestsService } from './requests.service';
 import { PermissionsService } from './permissions.service';
