+++ src/app/iot/locations/locations.component.spec.ts
@@ -6,7 +6,7 @@ import { RequestsService } from '@app/core/services/requests.service';
 import { LocationsComponent } from './locations.component';
 import { PlacesComponent } from './../index/places/places.component';
 import { CommunicateService } from '@app/core/services/communicate.service';
-import { appReducersGenerator } from '@app/core/app.reducers'; ;
+import { appReducersGenerator } from '@app/app.reducers'; ;
 import { ActivatedRoute } from '@angular/router';
 import { PermissionsService } from '@app/core/services/permissions.service';
 import { MocksService } from '@app/core/services/mocks.service';
