+++ src/app/iot/shared/helper-box/helper-box.component.spec.ts
@@ -3,7 +3,7 @@ import { HelperBoxComponent } from './helper-box.component';
 import { OutputPinComponent } from '../output-pin/output-pin.component';
 import { SelectInputComponent } from '@app/core/forms/select-input/select-input.component';
 import { TextInputComponent } from '@app/core/forms/text-input/text-input.component';
-import { appReducersGenerator } from '@app/core/app.reducers';
+import { appReducersGenerator } from '@app/app.reducers';
 
 import { CommunicateService } from '@app/core/services/communicate.service';
 import { RequestsService } from '@app/core/services/requests.service';
